Fix failing nix build
After upgrading to latest version in nixpkgs, some Go tests started
failing, beacuse nix enforces a sandbox testing environment, which
fails tests that rely on implicit dependencies.

For this to not happen again, we now run `nix build` in our unit test
CI action, so that Go tests are also run in a sandboxed environment.